+++ src/main/java/de/retest/web/screenshot/ScreenshotProvider.java
@@ -7,8 +7,8 @@ import org.openqa.selenium.WebDriver;
 /**
  * The screenshot provider for way screenshots are taken.
  *
- * The ScreenshotProvider can choose whether screenshots are taken for the full page or just the viewport. The Default
- * implementation is {@link ViewportOnlyScreenshot}.
+ * The ScreenshotProvider can choose whether screenshots are taken for the full page, just the viewport or none at all.
+ * The Default implementation is {@link ViewportOnlyScreenshot}.
  */
 public interface ScreenshotProvider {