+++ CHANGELOG.md
@@ -25,13 +25,14 @@ Table of Contents
 
 ### Breaking Changes
 
+* The default of how screenshots are created has changed from full page to viewport only. These two options are now available to set globally via properties and per test individually via `RecheckWebOptions`.
+
 ### Bug Fixes
 
 ### New Features
 
 ### Improvements
 
-* The default of how screenshots are created has changed from full page to viewport only. These two options are now available to set globally via properties and per test individually via `RecheckWebOptions`.
 
 
 --------------------------------------------------------------------------------